fix(switch): remove X overflow of the Switch internals
When positioned on the right with flex-row-reverse, the internal
components of the Switch caused overflow.
When on the right side of a page, this caused overflow of the content and
a horizontal scroll outside of the `html` element.
The x-overflow-hidden is not conditional because it has no effect on the basic layout.